+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Date format: DD/MM/YYYY
 - `DisableAcrylic` now fully disable transparency of its decendents `Acrylic`s ([#468](https://github.com/bdlukaa/fluent_ui/issues/468))
 - `Combobox.comboboxColor` is now correctly applied ([#468](https://github.com/bdlukaa/fluent_ui/issues/468))
 - Use the correct color for `DefaultToggleSwitchThumb` ([#468](https://github.com/bdlukaa/fluent_ui/issues/463)) 
+- Do not interpolate between infinite constraints on `TabView` ([#430](https://github.com/bdlukaa/fluent_ui/issues/430))
 
 ## [4.0.0-pre.3] - Top navigation and auto suggestions - [13/08/2022]